% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-imports.R
\docType{package}
\name{IEOstats-package}
\alias{IEOstats}
\alias{IEOstats-package}
\title{Internal imports of the package}
\description{
Statistical methods and tools for scientific data analysis.
}
\author{
\strong{Maintainer}: Joel Morales \email{jmarcelo.morales@ieo.csic.es}

Authors:
\itemize{
  \item Joel Morales \email{jmarcelo.morales@ieo.csic.es}
  \item Adriana Nogueira \email{adriana.nogueira@ieo.csic.es}
  \item Lizeth Rubio
}

}
\keyword{internal}
